Extract install controller into controllers/install.rs
