+++ slint-ui/src/controllers/mod.rs
@@ -1,1 +1,2 @@
+pub mod install;
 pub mod welcome;
